sched/tune: Only enqueue boosted tasks
Due to how we filter enqueued and non-enqueued tasks, create a separate function that will not run the said checks including the input timeout. This shall prevent non-boosted tasks from being tracked unnecessarily by boostgroups which do not have a boost or those that does have a boost but restricts it only for certain tasks such as top-app cgroup. Rearrange the functions prior to enqueue function.

Signed-off-by: ederekun <sedrickvince@gmail.com>