sched/tune: Introduce schedtune_boost_max and store actual number of SchedTune CGroups
This will allow us to have a copy of the highest boost value amongst all CGroups. Given this, in order to optimize the loops and avoid checking if allocate_group is NULL every time, let's simply assign a variable which will hold the value of the actual boost groups being utilized by the system. Use the allocated_group_num_idx also for schedtune_cpu_update.

Signed-off-by: ederekun <sedrickvince@gmail.com>