sched/tune: Dedicate a fast path for maximum priority tasks to passthrough
As of now, we need to go through lengthy RCU locks and BG locks just to manage the boostgroups - that is not really ideal for tasks we need to get finished as soon as possible. And so, by using adj values and a dedicated atomic_t counter for accounting these max prio tasks, let's provide a completely alternative pathway in all schedtune handles to bypass the boostgroup pathway and instead have the tasks be enqueued atomically within a separate variable. This separate variable will then be read first and reflect the maximum boost value recorded, same goes with task-specific schedtune parameters. We also retain the adj bypass behavior reflected by ST_LOW_PRIO, meaning that those of ST_HIGH_PRIO will only be the ones to go through the boostgroup pathway and be evaluated according to their SchedTune CGroup.

Signed-off-by: ederekun <sedrickvince@gmail.com>