cpufreq: schedutil: Introduce schedutil_interactive_timeout
This timeout follows the same structure as schedtune_input_timeout. The main purpose of this timeout is to decide on whether to invert the rate limits of schedutil which is currently configured for interactivity. Having it reverted produces the opposite effect (uninteractive responsiveness), leading to great battery savings but a severe degradation in performance; such is why we base the timestamp update on the framebuffer to prevent frame drops which is of most importance when it comes to interactivity.

Signed-off-by: ederekun <sedrickvince@gmail.com>