lwtimeout: Introduce Lightweight Timeout
This header provides a small suite of functions which aims to declare a set of static functions capable of providing lockless read of a variable that indicates a timer timeout, all the while having multiple threads updating the timer concurrently which may induce data races if done normally with a timer. Not only does this eliminate dependence towards a timer, it also allows the programmer to define at which code paths should the timeout be checked to further limit races. Do take note that this sacrifices some accuracy at the benefit of faster code execution.

Within this commit, also adapt the following timeout-dependent features:
	- SchedTune Input
	- Schedutil Interactive

Signed-off-by: ederekun <sedrickvince@gmail.com>