sched/tune: Rename SchedTune Input to SchedTune Interactive
To unify and simplify nomenclature, have SchedTune Input be instead defined as SchedTune Interactive which matches Schedutil Interactive. From this moment on as well, they also now become a part of Sched Interactive at which manages the responsiveness of the scheduler depending on how userspace interacts with the device. Along these changes, we also adjust the update hooks of sched interactive to keep away from hot code paths.

Signed-off-by: ederekun <sedrickvince@gmail.com>