sched/tune: Remap and introduce raw spinlocks for task management and modify priority assignments
Prevent data races for the max priority task code path by matching the lock convention used for boostgroups. Keep the locks short and comprehensive to reduce the performance overhead. In addition to this, let's also fix the logic of allowing RT tasks to be placed in maximum priority, instead we make RT tasks that are not of adj value less than 200 go through the normal boostgroup paths wherein they are boosted according to their CGroup.

In addition to this, separate the lock from the boost group structure to avoid contention with other variables within the structure when they are accessed locklessly.

Signed-off-by: ederekun <sedrickvince@gmail.com>