sched/tune: No longer filter high priority tasks
Given that we already filter by adj value and policy type, there is no longer need to filter by boost value as well as it is excessive and may cause boost to dip during a certain amount of time. Remove the boost check and restore the schedtune_task_boost function.

Signed-off-by: ederekun <sedrickvince@gmail.com>