sched/tune: Reduce the frequency at which lwtimeout updates
Synchronize the timeout updates with schedtune's boostgroup updates, prevent updating prior to checks as there may be in some cases that the timestamp CPU cache may not have yet synchronized with an update that had occured parallel to the update which causes unwanted nullifying of boosts during interactive periods. In addition to this, rearrange priority conditions in boostgroup enqueue to match the order of the schedtune parameter handlers.

Signed-off-by: ederekun <sedrickvince@gmail.com>