input: synaptics_driver_s3320: Extend compatibility
With the introduction of AOSPA's backported s3320 driver, a different configuration of the screen-off gestures has been again introduced. This patch extends compatibility to support the backport and allow ROMs that handle the gesture 'switch' in userspace be supported with their alternative key codes. Given that it's still dependent on the gesture 'enable' node, the node now signifies the use for alternative key codes - this is safe as OOS do not utilize their own key codes (only custom ROMs), and that with its exclusive use of gesture 'switch' node all that is needed is to bypass the key code selection and just set it to KEY_F4.

Signed-off-by: ederekun <sedrickvince@gmail.com>